requirements.txt
# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Computer Vision & Hand Tracking
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Voice Recognition
openai-whisper>=20231117
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Language Model Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Data Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0 